chore: git is not detecting file renames
